backend/requirements.txt
# Web framework
fastapi>=0.115.0,<0.117
uvicorn[standard]>=0.30.6,<0.35
gunicorn>=22.0.0

# Validation + settings — 2.10+ ships prebuilt wheels for Python 3.13 (Windows/macOS/Linux)
pydantic>=2.10.4,<3
pydantic-settings>=2.6.1,<3
email-validator>=2.2.0

# Database
SQLAlchemy>=2.0.35,<2.1

# Auth
python-jose[cryptography]>=3.3.0,<4
passlib==1.7.4

# Multipart uploads / forms
python-multipart>=0.0.12

# HTTP client (for any external API calls, e.g. Razorpay, AI providers)
httpx>=0.27.2,<0.29

# Razorpay (payments, optional at runtime)
razorpay>=1.4.2
